Separated Amps task and app.
